feat: modernise CoolerController with secrets, CI, bug fixes, EEPROM, compatibility
Agent-Logs-Url: https://github.com/bnorth12/CoolerController/sessions/f67be4de-5a22-42ba-90f8-6bf596e13137

+++ Timing.ino
@@ -1,7 +1,6 @@
 void timing() {
-  // put your main code here, to run repeatedly:
 
-  int currentMillis = millis();
+  unsigned long currentMillis = millis();
 
   if ((currentMillis - previousRG1 > rg1)) //Check RG1
   {
